# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Perk Up Daily API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: number;
  email: string;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  fullName?: string | null;
  /** @nullable */
  profilePicture?: string | null;
  /** @nullable */
  phone?: string | null;
  isAdmin?: boolean;
  isPremium?: boolean;
  /** @nullable */
  birthday?: string | null;
  onboardingCompleted?: boolean;
  createdAt?: string;
}

export interface AuthResponse {
  token: string;
  user: AuthUser;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface RegisterInput {
  email: string;
  password: string;
  firstName?: string;
  lastName?: string;
}

export interface UserUpdate {
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  fullName?: string | null;
  /** @nullable */
  profilePicture?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  birthday?: string | null;
  onboardingCompleted?: boolean;
  isPremium?: boolean;
  dailyCheckInEnabled?: boolean;
  notificationsEnabled?: boolean;
}

export interface ForgotPasswordInput {
  email: string;
}

export interface ResetPasswordInput {
  resetToken: string;
  newPassword: string;
}

/**
 * @nullable
 */
export type UserEntryMetadata = { [key: string]: unknown } | null;

export interface UserEntry {
  id: number;
  userId: number;
  entryType: string;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  metadata?: UserEntryMetadata;
  createdAt: string;
  updatedAt?: string;
}

export type UserEntryInputMetadata = { [key: string]: unknown };

export interface UserEntryInput {
  entryType: string;
  title?: string;
  content?: string;
  metadata?: UserEntryInputMetadata;
}

export type UserEntryUpdateMetadata = { [key: string]: unknown };

export interface UserEntryUpdate {
  title?: string;
  content?: string;
  metadata?: UserEntryUpdateMetadata;
}

export interface LibraryItem {
  id: number;
  contentType: string;
  /** @nullable */
  title?: string | null;
  content: string;
  /** @nullable */
  author?: string | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  tags?: string[] | null;
  status: string;
  sortOrder?: number;
  /** @nullable */
  imageUrl?: string | null;
  createdAt?: string;
}

export interface LibraryItemInput {
  contentType: string;
  title?: string;
  content: string;
  author?: string;
  category?: string;
  status?: string;
  sortOrder?: number;
  imageUrl?: string;
}

export interface LibraryItemUpdate {
  title?: string;
  content?: string;
  author?: string;
  category?: string;
  status?: string;
  sortOrder?: number;
  imageUrl?: string;
}

export interface LibraryBulkInput {
  items: LibraryItemInput[];
}

export interface ReflectionPrompt {
  id: number;
  prompt: string;
  /** @nullable */
  category?: string | null;
  status: string;
  createdAt?: string;
}

export interface ReflectionPromptInput {
  prompt: string;
  category?: string;
  status?: string;
}

export interface ReflectionPromptUpdate {
  prompt?: string;
  category?: string;
  status?: string;
}

export interface BugReport {
  id: number;
  /** @nullable */
  userId?: number | null;
  description: string;
  status: string;
  /** @nullable */
  adminNotes?: string | null;
  createdAt?: string;
}

export interface BugReportInput {
  description: string;
}

export interface BugReportUpdate {
  status?: string;
  adminNotes?: string;
}

/**
 * @nullable
 */
export type NeurocycleCheckInStepData = { [key: string]: unknown } | null;

export interface NeurocycleCheckIn {
  id: number;
  userId: number;
  day: number;
  /** @nullable */
  stepData?: NeurocycleCheckInStepData;
  /** @nullable */
  completedAt?: string | null;
  createdAt?: string;
}

export type NeurocycleCheckInInputStepData = { [key: string]: unknown };

export interface NeurocycleCheckInInput {
  day: number;
  stepData?: NeurocycleCheckInInputStepData;
  completedAt?: string;
}

export type NeurocycleCheckInUpdateStepData = { [key: string]: unknown };

export interface NeurocycleCheckInUpdate {
  day?: number;
  stepData?: NeurocycleCheckInUpdateStepData;
  completedAt?: string;
}

export interface NeuralTrainingItem {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  exerciseType: string;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  videoUrl?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  durationMinutes?: number | null;
  status: string;
  sortOrder?: number;
  createdAt?: string;
}

export type ListEntriesParams = {
entry_type?: string;
limit?: number;
};

export type ListLibraryParams = {
content_type?: string;
status?: string;
limit?: number;
};

export type ListPromptsParams = {
status?: string;
};

export type ListBugReportsParams = {
status?: string;
limit?: number;
};

export type ListNeuralTrainingParams = {
exercise_type?: string;
status?: string;
};

export type ListUsersParams = {
limit?: number;
};