# lib/api-zod/src/generated/types/bugReport.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Perk Up Daily API
 * OpenAPI spec version: 0.1.0
 */

export interface BugReport {
  id: number;
  /** @nullable */
  userId?: number | null;
  description: string;
  status: string;
  /** @nullable */
  adminNotes?: string | null;
  createdAt?: string;
}