# lib/api-zod/src/generated/types/bugReportUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Perk Up Daily API
 * OpenAPI spec version: 0.1.0
 */

export interface BugReportUpdate {
  status?: string;
  adminNotes?: string;
}