# lib/api-zod/src/generated/types/libraryItemInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Perk Up Daily API
 * OpenAPI spec version: 0.1.0
 */

export interface LibraryItemInput {
  contentType: string;
  title?: string;
  content: string;
  author?: string;
  category?: string;
  status?: string;
  sortOrder?: number;
  imageUrl?: string;
}