# lib/api-zod/src/generated/types/userEntry.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Perk Up Daily API
 * OpenAPI spec version: 0.1.0
 */
import type { UserEntryMetadata } from './userEntryMetadata';

export interface UserEntry {
  id: number;
  userId: number;
  entryType: string;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  metadata?: UserEntryMetadata;
  createdAt: string;
  updatedAt?: string;
}